requirements.txt
Django==2
djangorestframework
Pillow
drfapikey
django-cors-headers
channels
channels_redis
psycopg2
dj_database_url
django-extensions
